refactor: 수정사항 모두 리펙토링 완료했습니다

+++ hugme-dog/src/main/java/spring/hugme/domain/community/dto/BoardListResponse.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.UUID;
 import lombok.Builder;
 import lombok.Data;
-import spring.hugme.domain.community.code.BoardAlias;
+import spring.hugme.global.code.BoardAlias;
 
 @Data
 @Builder
@@ -26,9 +26,9 @@ public class BoardListResponse {
 
   String content;
 
-  int likeCount;
+  Long likeCount;
 
-  int commentCount;
+  Long commentCount;
 
 
 }
